+++ R/control_survival.R
@@ -30,14 +30,14 @@ control_coxph <- function(pval_method = c("log-rank", "wald", "likelihood"),
 #'
 #' @md
 #' @inheritParams argument_convention
-#' @param conf_type (`string`) \cr "plain" (default), "none", "log", "log-log" for confidence interval type, \cr
-#'    see more in [survival::survfit()]
+#' @param conf_type (`string`) \cr "plain" (default), "log", "log-log" for confidence interval type, \cr
+#'    see more in [survival::survfit()]. Note that the option "none" is no longer supported.
 #' @param quantiles (`numeric`) \cr of length two to specify the quantiles of survival time.
 #' @return A list of components with the same names as the arguments
 #' @export
 #'
 control_surv_time <- function(conf_level = 0.95,
-                              conf_type = c("plain", "none", "log", "log-log"),
+                              conf_type = c("plain", "log", "log-log"),
                               quantiles = c(0.25, 0.75)) {
   conf_type <- match.arg(conf_type)
   assert_that(
@@ -60,7 +60,7 @@ control_surv_time <- function(conf_level = 0.95,
 #' @export
 #'
 control_surv_timepoint <- function(conf_level = 0.95,
-                                   conf_type = c("plain", "none", "log", "log-log")
+                                   conf_type = c("plain", "log", "log-log")
 ) {
   conf_type <- match.arg(conf_type)
   assert_that(is_proportion(conf_level))