+++ R/survival_timepoint.R
@@ -8,8 +8,8 @@
 #' @param control a (`list`) of parameters for comparison details, specified by using \cr
 #'    the helper function [control_surv_timepoint]. Some possible parameter options are: \cr
 #' * `conf_level`: (`proportion`)\cr confidence level of the interval for survival rate.
-#' * `conf_type`: (`string`) \cr "plain" (default), "none", "log", "log-log" for confidence interval type, \cr
-#'    see more in [survival::survfit()]
+#' * `conf_type`: (`string`) \cr "plain" (default), "log", "log-log" for confidence interval type, \cr
+#'    see more in [survival::survfit()]. Note that the option "none" is no longer supported.
 #' * `time_point`: (`number`) \cr survival time point of interest.
 #'
 #' @name survival_timepoint